payMoney int 改为float

+++ report_receive_data/src/main/java/com/seastar/entity/UserPayReq.java
@@ -8,6 +8,6 @@ import java.util.Date;
 public class UserPayReq extends BaseRequest
 {
     public long userId;
-    public int payMoney;
+    public float payMoney;
     public Date serverTime;
 }
